fix commit transactionController
